Created Repository interface
