+++ src/main/java/unq/dapp/supergol/model/repositories/Repository.java
@@ -1,28 +1,15 @@
 package unq.dapp.supergol.model.repositories;
 
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Optional;
 import java.util.function.Predicate;
-import java.util.stream.Collectors;
 
-public class Repository<T extends Persistable> {
-  private Collection<T> elements = new ArrayList<T>();
+public interface Repository<T extends Persistable> {
+  Collection<T> all();
 
-  public Collection<T> all() {
-    return elements;
-  }
+  Optional<T> findById(long id);
 
-  public Optional<T> findById(long id) {
-    return elements.stream().filter(it -> it.getId() == id).findAny();
-  }
+  Collection<T> filter(Predicate<T> comparator);
 
-  public Collection<T> filter(Predicate<T> comparator) {
-    return elements.stream().filter(comparator).collect(Collectors.<T>toList());
-  }
-
-  public Repository<T> add(T element) {
-    elements.add(element);
-    return this;
-  }
+  Repository<T> add(T element);
 }
