+++ src/test/java/unq/dapp/supergol/model/StageImportTest.java
@@ -3,6 +3,7 @@ package unq.dapp.supergol.model;
 import org.junit.Before;
 import org.junit.Test;
 import unq.dapp.supergol.model.exceptions.UnexistentPlayerException;
+import unq.dapp.supergol.model.repositories.CollectionBasedRepository;
 import unq.dapp.supergol.model.repositories.Repository;
 
 import java.time.LocalDate;
@@ -18,7 +19,7 @@ public class StageImportTest {
 
   @Before
   public void setUp() {
-    repo = new Repository<>();
+    repo = new CollectionBasedRepository<>();
 
     RealWorldTeam racing = anyRealWorldTeam();
     milito = Player.forward(racing);
